% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-harmonics.R
\name{format.tide_harmonics}
\alias{format.tide_harmonics}
\title{Format tide_harmonics}
\usage{
\method{format}{tide_harmonics}(x)
}
\arguments{
\item{x}{The object to format.}
}
\value{
A list of the tide_harmonics components.
}
\description{
Formats a tide_harmonics object.
}

