% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonics}
\alias{harmonics}
\title{Harmonics}
\format{
An object of class \code{tide_harmonics} of length 4.
}
\usage{
harmonics
}
\description{
A object of class tide_harmonics providing tidal harmonic data for US stations.
}
\details{
Converted from harmonics-dwf-20261228-free.tar.xz, NOAA web site data processed by David Flater for XTide.
}
\keyword{datasets}
