% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-harmonics.R
\name{is.tide_harmonics}
\alias{is.tide_harmonics}
\title{Is tide_harmonics}
\usage{
is.tide_harmonics(x)
}
\arguments{
\item{x}{The object to test.}
}
\description{
Tests if object inherits from class tide_harmonics.
}

