% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monterey}
\alias{monterey}
\title{Monterey Tide Height Data}
\format{A tbl data frame:
\describe{
  \item{DateTime}{The date time (time).}
  \item{MLLW}{The tide height in m (dbl).}
}}
\usage{
monterey
}
\description{
High/Low Tide Predictions for Monterey (StationID: 9413450).
}
\details{
The predictions were taken from \url{http://tidesandcurrents.noaa.gov/noaatidepredictions/NOAATidesFacade.jsp?Stationid=9413450}.
}
\keyword{datasets}

