% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{noaa}
\alias{noaa}
\title{noaa}
\format{An object of class \code{rtide} of length 4.}
\usage{
noaa
}
\description{
A object of class rtide providing harmonics and offsets for NOAA tide stations.
The information was scraped from \url{"https://tidesandcurrents.noaa.gov"} on April 5, 2017.
}
\details{
Please read the disclaimer at \url{https://tidesandcurrents.noaa.gov/disclaimers.html}.
}
\keyword{datasets}
