% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-rtide.R
\name{predict_rtide}
\alias{predict_rtide}
\title{Predict Tide Height}
\usage{
predict_rtide(data, rtide = rtide::noaa, slack = FALSE, ...)
}
\arguments{
\item{data}{A data.frame with the columns DateTime and Station.}

\item{rtide}{The rtide object to use for the predictions.}

\item{slack}{A flag indicating whether to also calculate the time and height of the closest slack tide.}

\item{...}{Unused arguments.}
}
\value{
An updated data.frame with the additional column TideHeight and if \code{slack = TRUE}
the additional columns DateTimeSlack and TideHeightSlack.
}
\description{
Predicts tide heights (in m) at stations and date times provided in new_data.
}
