% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-slack-tide.R
\name{predict_slack_tide}
\alias{predict_slack_tide}
\title{Predict Tide Height}
\usage{
predict_slack_tide(data, rtide = rtide::noaa, ...)
}
\arguments{
\item{data}{A data.frame with the columns DateTime and Station.}

\item{rtide}{The rtide object to use for the predictions.}

\item{...}{Unused arguments.}
}
\value{
An updated data.frame with the additional columns SlackDateTime, SlackTideHeight, SlackType.
}
\description{
Predicts closest slack tide time and height (in m) at stations and date times provided in new_data.
}
