% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict-tide-height.R
\name{predict_tide_height}
\alias{predict_tide_height}
\title{Predict Tide Height}
\usage{
predict_tide_height(data, rtide = rtide::noaa, ...)
}
\arguments{
\item{data}{A data.frame with the columns DateTime and Station.}

\item{rtide}{The rtide object to use for the predictions.}

\item{...}{Unused arguments.}
}
\value{
An updated data.frame with the additional column TideHeight.
}
\description{
Predicts tide heights (in m) at stations and date times provided in new_data.
}
\details{
predict_tide_height is only defined for stations with harmonics which are indicated
by non-missing Datum values.
}
