% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtide-package.R
\docType{package}
\name{rtide-package}
\alias{rtide}
\alias{rtide-package}
\title{rtide: Tide Heights}
\description{
Calculates tide heights based on tide station harmonics. It includes the harmonics data for 1285 US stations. The harmonics data was converted from \url{https://github.com/millerlp/rtide/blob/main/data-raw/harmonics-dwf-20251228-free.tar.xz} NOAA web site data processed by David Flater for 'XTide 2'. The code to calculate tide heights from the harmonics is based on 'XTide 2'.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/millerlp/rtide}
  \item Report bugs at \url{https://github.com/millerlp/rtide/issues}
}

}
\author{
\strong{Maintainer}: Luke Miller \email{contact@lukemiller.org}

Authors:
\itemize{
  \item Luke Miller \email{contact@lukemiller.org}
  \item Joe Thorley \email{joe@poissonconsulting.ca} (\href{https://orcid.org/0000-0002-7683-4592}{ORCID})
  \item Abram Fleishman
}

Other contributors:
\itemize{
  \item Poisson Consulting [copyright holder]
}

}
\keyword{internal}
