% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{seq_datetime}
\alias{seq_datetime}
\title{Seq Date Times}
\usage{
seq_datetime(from = Sys.Date(), to = from, minutes = 360L, tz = "UTC")
}
\arguments{
\item{from}{A Date of the start of the period of interest}

\item{to}{A Date of the end of the period of interest}

\item{minutes}{An integer of the number of minutes between date times.}

\item{tz}{A string of the time zone.}
}
\value{
A POSIXct vector.
}
\description{
A helper function to generate a sequence of date times.
}
\examples{
seq_datetime()
}
