% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-height.R
\name{tide_datetimes}
\alias{tide_datetimes}
\title{Tide Date Times}
\usage{
tide_datetimes(
  minutes = 60L,
  from = as.Date("2015-01-01"),
  to = as.Date("2015-12-31"),
  tz = "PST8PDT"
)
}
\arguments{
\item{minutes}{An integer of the number of minutes between tide heights}

\item{from}{A Date of the start of the period of interest}

\item{to}{A Date of the end of the period of interest}

\item{tz}{A string of the time zone.}
}
\value{
A POSIXct vector.
}
\description{
Generates sequence of date times.
}
\examples{
tide_datetimes()
}
