% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-height.R
\name{tide_height_data}
\alias{tide_height_data}
\title{Tide Height Data}
\usage{
tide_height_data(data, harmonics = rtide::harmonics)
}
\arguments{
\item{data}{A data frame with the columns Station and DateTime.}

\item{harmonics}{The harmonics object.}
}
\value{
A data frame of the tide heights in meters.
}
\description{
Calculates tide height at specified stations at particular date times based on the supplied harmonics object.
}
