% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide.R
\name{tide_stations}
\alias{tide_stations}
\title{Tide Stations}
\usage{
tide_stations(stations = ".*", harmonics = rtide::harmonics)
}
\arguments{
\item{stations}{A character vector of stations to match - treated as regular expressions.}

\item{harmonics}{The harmonics object.}
}
\value{
A character vector of the matching stations in the harmonics object.
}
\description{
Tide Stations
}
\examples{
tide_stations()
}

