% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonics}
\alias{harmonics}
\title{Harmonics}
\format{An object of class \code{tide_harmonics} of length 4.}
\usage{
harmonics
}
\description{
A object of class tide_harmonics providing tidal harmonic data for US stations.
}
\details{
The tidal harmonic data was originally derived from XTide (http://www.flaterco.com/xtide/)
by Luke Miller (https://github.com/millerlp/Misc_R_scripts).
}
\keyword{datasets}

