% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonics}
\alias{harmonics}
\title{Harmonics}
\format{An object of class \code{list} of length 14.}
\usage{
harmonics
}
\description{
A list of harmonics data for US sites.
}
\keyword{datasets}

