% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{monterey}
\alias{monterey}
\title{Monterey Tide Height Data}
\format{A tbl data frame:
\describe{
\item{Station}{The station name (chr).}
\item{DateTime}{The date time (time).}
\item{MLLW}{The tide height in m (dbl).}
}}
\usage{
monterey
}
\description{
High/Low Tide Predictions from \url{http://tidesandcurrents.noaa.gov/tide_predictions.html}.
}
\keyword{datasets}
