% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-slack.R
\name{tide_slack_data}
\alias{tide_slack_data}
\title{Tide Slack Data}
\usage{
tide_slack_data(data, harmonics = rtide::harmonics)
}
\arguments{
\item{data}{A data frame with the columns Station and DateTime.}

\item{harmonics}{The harmonics object.}
}
\value{
A data frame of the slack tide date times and heights in m.
}
\description{
Determines the closest slack tide for specified stations at particular date times based on the supplied harmonics object.
}
