% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tide-height.R
\name{tide_stations}
\alias{tide_stations}
\title{Tide Stations}
\usage{
tide_stations(stations = ".*", harmonics = rtide::harmonics)
}
\arguments{
\item{stations}{A character vector of stations to match - treated as regular expressions.}

\item{harmonics}{The harmonics object.}
}
\description{
Gets vector of matching stations.
}

